Move sparkctl to cmd directory (#2347)
* Move spark-operator


* Move sparkctl to cmd directory


* Remove unnecessary app package/directory


---------

(cherry picked from commit 2375a306f9c0e85fb486ef58aa6d5aebba8dc8b2)
