+++ cmd/sparkctl/app/testdata/hadoop-conf/core-site.xml
