+++ cmd/sparkctl/app/utils.go
@@ -1,11 +1,11 @@
 /*
-Copyright 2024 The Kubeflow authors.
+Copyright 2018 Google LLC
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
 
-    http://www.apache.org/licenses/LICENSE-2.0
+    https://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,18 +14,26 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package app
 
 import (
-	"fmt"
-	"os"
+	"time"
 
-	"github.com/kubeflow/spark-operator/cmd/operator"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/duration"
 )
 
-func main() {
-	if err := operator.NewCommand().Execute(); err != nil {
-		fmt.Fprintf(os.Stderr, "%v\n", err)
-		os.Exit(1)
+func getSinceTime(timestamp metav1.Time) string {
+	if timestamp.IsZero() {
+		return "N.A."
 	}
+
+	return duration.ShortHumanDuration(time.Since(timestamp.Time))
+}
+
+func formatNotAvailable(info string) string {
+	if info == "" {
+		return "N.A."
+	}
+	return info
 }
